feature-infos - reorg/rename to better document
