+++ projects/eav-ui/src/app/app-administration/app-configuration/app-configuration.html
@@ -408,7 +408,7 @@
       </mat-card-header>
       <mat-card-content>
         <p>LightSpeed Output Cache to speed up page by 50-90%</p>
-        <app-feature-text-info [featureNameId]="LightSpeedOutputCache" [asInfo]="true"></app-feature-text-info>
+        <app-feature-info-box [featureNameId]="LightSpeedOutputCache" [asInfo]="true"></app-feature-info-box>
       </mat-card-content>
       <div class="eav-card-space-div"></div>
       <mat-card-actions align="start">
@@ -440,7 +440,7 @@
       </mat-card-header>
       <mat-card-content>
         <p>ContentSecurityPolicy (CSP) features in this App - for maximum control.</p>
-        <app-feature-text-info [featureNameId]="ContentSecurityPolicy" [asInfo]="true"></app-feature-text-info>
+        <app-feature-info-box [featureNameId]="ContentSecurityPolicy" [asInfo]="true"></app-feature-info-box>
       </mat-card-content>
       <div class="eav-card-space-div"></div>
       <mat-card-actions align="start">
