+++ projects/eav-ui/src/app/app-administration/import-export-menu/app-state/app-state.html
@@ -28,7 +28,7 @@
     <br />
     They will be placed in a folder inside <code>App_Data</code> and allow sync/restore with git or another server.
   </p>
-  <app-feature-text-info [featureNameId]="AppSyncWithSiteFiles"></app-feature-text-info>
+  <app-feature-info-box [featureNameId]="AppSyncWithSiteFiles"></app-feature-info-box>
 
     <mat-dialog-actions class="eav-dialog-actions" style="height: 73px">
       <div class="eav-dialog-actions-div">
