+++ projects/eav-ui/src/app/app-administration/import-export-menu/app-state/app-state.ts
@@ -5,8 +5,8 @@ import { MatDialogActions } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { transient } from '../../../../../../core';
+import { FeatureTextInfoComponent } from '../../../features/feature-info-box/feature-info-box';
 import { FeatureNames } from '../../../features/feature-names';
-import { FeatureTextInfoComponent } from '../../../features/feature-text-info/feature-text-info';
 import { FeaturesService } from '../../../features/features.service';
 import { ExportAppService } from '../../services/export-app.service';
 import { ImportAppPartsService } from '../../services/import-app-parts.service';
