+++ projects/eav-ui/src/app/app-administration/import-export-menu/data-bundles/data-bundles.html
@@ -6,7 +6,7 @@
     a partial sync of data and schema between environments.
   </p>
 
-  <app-feature-text-info [featureNameId]="DataExportImportBundles"></app-feature-text-info>
+  <app-feature-info-box [featureNameId]="DataExportImportBundles"></app-feature-info-box>
 
   <!-- Explanation about other places to export/import -->
   <div class="info-container">
