+++ projects/eav-ui/src/app/app-administration/import-export-menu/data-bundles/data-bundles.ts
@@ -11,8 +11,8 @@ import { EntityLightIdentifier } from 'projects/edit-types/src/EntityLight';
 import { take } from 'rxjs';
 import { ContentItem } from '../../../content-items/models/content-item.model';
 import { ContentItemsService } from '../../../content-items/services/content-items.service';
+import { FeatureTextInfoComponent } from '../../../features/feature-info-box/feature-info-box';
 import { FeatureNames } from '../../../features/feature-names';
-import { FeatureTextInfoComponent } from '../../../features/feature-text-info/feature-text-info';
 import { FeaturesService } from '../../../features/features.service';
 import { AgGridHelper } from '../../../shared/ag-grid/ag-grid-helper';
 import { ColumnDefinitions } from '../../../shared/ag-grid/column-definitions';
