+++ projects/eav-ui/src/app/app-administration/import-export-menu/export-app/export-app.ts
@@ -6,9 +6,9 @@ import { MatDialogActions } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { transient } from '../../../../../../core';
 import { DocsLinkHelper } from '../../../admin-shared/docs-link-helper/docs-link-helper';
-import { FeatureIconComponent } from '../../../features/feature-icon/feature-icon';
 import { FeatureNames } from '../../../features/feature-names';
 import { FeaturesService } from '../../../features/features.service';
+import { FeatureIconWithDialogComponent } from '../../../features/icons/feature-icon-with-dialog';
 import { ExportAppService } from '../../services/export-app.service';
 
 @Component({
@@ -21,7 +21,7 @@ import { ExportAppService } from '../../services/export-app.service';
         MatDialogActions,
         MatButtonModule,
         MatIconModule,
-        FeatureIconComponent,
+        FeatureIconWithDialogComponent,
         DocsLinkHelper,
     ]
 })
