+++ projects/eav-ui/src/app/app-administration/recycle-bin/recycle-bin.html
@@ -12,7 +12,7 @@
       Deleted Items ({{ deletedEntities().length }} total)
 
       <!-- Indicator if feature is not active -->
-      <app-feature-icon [featureNameId]="'EntityUndelete'"></app-feature-icon>
+      <app-feature-icon-with-dialog [featureNameId]="'EntityUndelete'"></app-feature-icon-with-dialog>
     </h3>
     
     <div class="recycle-bin-filters">
