+++ projects/eav-ui/src/app/apps-management/add-app-from-folder/add-app-from-folder.html
@@ -3,7 +3,7 @@
     <div class="eav-dialog-header__title">Add App from Folder</div>
   </div>
   <div class="eav-dialog-content">
-    <app-feature-text-info [featureNameId]="AppSyncWithSiteFiles"></app-feature-text-info>
+    <app-feature-info-box [featureNameId]="AppSyncWithSiteFiles"></app-feature-info-box>
     <p>
       Select an app from the apps-folder which has been copied there (or git-cloned) but not installed yet.
     </p>
