+++ projects/eav-ui/src/app/apps-management/add-app-from-folder/add-app-from-folder.ts
@@ -4,8 +4,8 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatDialogActions, MatDialogRef } from "@angular/material/dialog";
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { transient } from '../../../../../core';
+import { FeatureTextInfoComponent } from '../../features/feature-info-box/feature-info-box';
 import { FeatureNames } from '../../features/feature-names';
-import { FeatureTextInfoComponent } from '../../features/feature-text-info/feature-text-info';
 import { FeaturesService } from '../../features/features.service';
 import { ColumnDefinitions } from '../../shared/ag-grid/column-definitions';
 import { IdFieldParams } from '../../shared/components/id-field/id-field.models';
