+++ projects/eav-ui/src/app/apps-management/system-info/system-info.html
@@ -87,7 +87,7 @@
       <mat-card-subtitle>Performance Enhancements and Caching</mat-card-subtitle>
     </mat-card-header>
     <mat-card-content>
-      <app-feature-text-info [featureNameId]="LightSpeedOutputCache" [asInfo]="true"></app-feature-text-info>
+      <app-feature-info-box [featureNameId]="LightSpeedOutputCache" [asInfo]="true"></app-feature-info-box>
       @if (lsEnabled()) {
         <div>LightSpeed is enabled. Configure it in each site/app as you need.</div>
       }
@@ -107,7 +107,7 @@
       <mat-card-subtitle>System Security Features</mat-card-subtitle>
     </mat-card-header>
     <mat-card-content>
-      <app-feature-text-info [featureNameId]="ContentSecurityPolicy" [asInfo]="true"></app-feature-text-info>
+      <app-feature-info-box [featureNameId]="ContentSecurityPolicy" [asInfo]="true"></app-feature-info-box>
       @if (cspEnabled()) {
         <div>ContentSecurityPolicy is enabled. Configure it in each site/app as you need.</div>
       }
