+++ projects/eav-ui/src/app/apps-management/system-info/system-info.ts
@@ -11,8 +11,8 @@ import { map, take } from 'rxjs';
 import { transient } from '../../../../../core';
 import { DocsLinkHelper } from '../../admin-shared/docs-link-helper/docs-link-helper';
 import { DialogConfigAppService } from '../../app-administration/services/dialog-config-app.service';
+import { FeatureTextInfoComponent } from '../../features/feature-info-box/feature-info-box';
 import { FeatureNames } from '../../features/feature-names';
-import { FeatureTextInfoComponent } from '../../features/feature-text-info/feature-text-info';
 import { FeaturesService } from '../../features/features.service';
 import { FieldHintComponent } from '../../shared/components/field-hint/field-hint';
 import { TippyDirective } from '../../shared/directives/tippy.directive';
