+++ projects/eav-ui/src/app/content-type-fields/field-sharing-add-many/field-sharing-add-many.html
@@ -101,9 +101,9 @@
     </div>
 
     <!-- Show feature warning -->
-    <app-feature-text-info [featureNameId]="ContentTypeFieldsReuseDefinitions" [asInfo]="true">
+    <app-feature-info-box [featureNameId]="ContentTypeFieldsReuseDefinitions" [asInfo]="true">
       <em> Even if the feature is not licensed, it will always work but show warnings. </em>
-    </app-feature-text-info>
+    </app-feature-info-box>
   </div>
   <!-- Save buttons -->
   <mat-dialog-actions align="end">
