+++ projects/eav-ui/src/app/edit/dialog/header/edit-dialog-header.ts
@@ -5,7 +5,7 @@ import { MatCardModule } from '@angular/material/card';
 import { MatDialog } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { TranslateModule } from '@ngx-translate/core';
-import { openFeaturesUsedButUnlicensedDialog } from '../../../features/features-used-but-missing-dialog/features-used-but-unlicensed-dialog';
+import { openFeaturesUsedButUnlicensedDialog } from '../../../features/dialogs/features-used-but-unlicensed-dialog';
 import { FeaturesService } from '../../../features/features.service';
 import { TippyDirective } from '../../../shared/directives/tippy.directive';
 import { FormConfigService } from '../../form/form-config.service';
