+++ projects/eav-ui/src/app/edit/fields/wrappers/feature-warning/feature-warning-wrapper.html
@@ -7,7 +7,7 @@
       This uses features which are not licensed:
       @for (feature of requiredFeatures(); track feature) {
         {{ feature }}
-        <app-feature-icon [featureNameId]="feature"></app-feature-icon>
+        <app-feature-icon-with-dialog [featureNameId]="feature"></app-feature-icon-with-dialog>
       }
     </div>
   }
