+++ projects/eav-ui/src/app/edit/fields/wrappers/hyperlink-dialog/hyperlink-default-expandable-wrapper.html
@@ -134,7 +134,7 @@
           @if (buttonAdam()) {
             <div class="c-ew__preview-field-hint c__helper-wrapper--right">
               <span>{{ adamSponsorI18nKey() | translate }}</span>
-              <app-feature-icon-text [featureNameId]="'NoSponsoredByToSic'"></app-feature-icon-text>
+              <app-feature-icon-with-dialog [forText]="true" [featureNameId]="'NoSponsoredByToSic'"></app-feature-icon-with-dialog>
             </div>
           }
         </div>
