+++ projects/eav-ui/src/app/edit/fields/wrappers/hyperlink-dialog/hyperlink-default-expandable-wrapper.ts
@@ -6,7 +6,7 @@ import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatIconModule } from "@angular/material/icon";
 import { MatInputModule } from "@angular/material/input";
 import { TranslateModule } from "@ngx-translate/core";
-import { FeatureIconTextComponent } from '../../../../features/feature-icon-text/feature-icon-text';
+import { FeatureIconWithDialogComponent } from 'projects/eav-ui/src/app/features/icons/feature-icon-with-dialog';
 import { FeatureNames } from '../../../../features/feature-names';
 import { FeaturesService } from '../../../../features/features.service';
 import { TippyDirective } from '../../../../shared/directives/tippy.directive';
@@ -38,7 +38,7 @@ import { WrappersCatalog } from "../wrappers.constants";
         TippyDirective,
         DialogPopupComponent,
         FieldHelperTextComponent,
-        FeatureIconTextComponent,
+        FeatureIconWithDialogComponent,
         PasteClipboardImageDirective,
         CommonModule,
         MatFormFieldModule,
