+++ projects/eav-ui/src/app/edit/fields/wrappers/hyperlink-library-dialog/hyperlink-library-expandable-wrapper.html
@@ -70,7 +70,7 @@
     </div>
     <div class="c__helper-wrapper--right c-ew__preview-hint-text" fxHide fxShow.gt-xs>
       <span>{{ adamSponsorI18nKey() | translate }} </span>
-      <app-feature-icon-text [featureNameId]="'NoSponsoredByToSic'"></app-feature-icon-text>
+      <app-feature-icon-with-dialog [forText]="true" [featureNameId]="'NoSponsoredByToSic'"></app-feature-icon-with-dialog>
     </div>
   </div>
 </div>
