+++ projects/eav-ui/src/app/edit/localization/auto-translate-menu-dialog/auto-translate-menu-dialog.html
@@ -1,5 +1,5 @@
 <mat-card class="auto-translate-menu-dialog-card u-hide-scrollbar">
-  <app-feature-text-info class="feature-info" [featureNameId]="EditUiTranslateWithGoogle"></app-feature-text-info>
+  <app-feature-info-box class="feature-info" [featureNameId]="EditUiTranslateWithGoogle"></app-feature-info-box>
   <mat-card-header>
     <mat-card-title>
       {{ 'LangMenu.Dialog.AutoTranslateAll.Title' | translate }}
