+++ projects/eav-ui/src/app/features/dialogs/feature-info-dialog.ts
