--- projects/eav-ui/src/app/features/feature-icon-text/feature-icon-text.html
@@ -1,3 +1,0 @@
-@if (show()) {
-  <mat-icon class="feature-icon-text" tippyTranslate="Features.NotActivated" [tippyData]="{ name: featureNameId() }" (click)="openDialog()">diamond</mat-icon>
-}
