--- projects/eav-ui/src/app/features/feature-icon-text/feature-icon-text.scss
@@ -1,3 +1,0 @@
-.feature-icon-text {
-  font-size: small;
-}
