+++ projects/eav-ui/src/app/features/feature-info-box/feature-info-box.scss
