+++ projects/eav-ui/src/app/features/feature-info-box/feature-info-box.ts
@@ -3,12 +3,12 @@ import { MatIconModule } from '@angular/material/icon';
 import { TranslateModule } from '@ngx-translate/core';
 import { TippyDirective } from '../../shared/directives/tippy.directive';
 import { computedObj } from '../../shared/signals/signal.utilities';
-import { FeatureComponentBase } from '../shared/base-feature';
+import { FeatureComponentBase } from '../shared/feature-component-base';
 
 @Component({
-  selector: 'app-feature-text-info',
-  templateUrl: './feature-text-info.html',
-  styleUrls: ['./feature-text-info.scss'],
+  selector: 'app-feature-info-box',
+  templateUrl: './feature-info-box.html',
+  styleUrls: ['./feature-info-box.scss'],
   imports: [
     MatIconModule,
     TranslateModule,
@@ -16,11 +16,14 @@ import { FeatureComponentBase } from '../shared/base-feature';
   ]
 })
 export class FeatureTextInfoComponent extends FeatureComponentBase {
+
+  /**
+   * Just indicate that this is an info, not a warning - so it can be styled differently if needed.
+   */
   asInfo = input<boolean>(false);
 
-  constructor() {
-    super();
-  }
-
+  /**
+   * Change the icon based on if this is info or warning, so it can be styled differently if needed.
+   */
   icon = computedObj('icon', () => this.asInfo() ? 'info' : 'warning');
 }
