+++ projects/eav-ui/src/app/features/icons/feature-icon-indicator.html
