+++ projects/eav-ui/src/app/features/icons/feature-icon-with-dialog.scss
@@ -1,0 +1,6 @@
+.feature-icon-text {
+  font-size: small;
+  width: 12px;
+  height: 12px;
+  padding: 0 4px;
+}
